lib/gstreamer: add common vpp base
Add a vpp base class for gstreamer to share
common code between msdk and vaapi elements.
